// src/types/masters.ts
export interface Department {
  id: string;
  code: string;
  name: string;
  created_at: string;
}

export interface Section {
  id: string;
  code: string;
  name: string;
  created_at: string;
}

export interface Role {
  id: string;
  code: string;
  name: string;
  created_at: string;
}

export interface DrawingCategory {
  id: string;
  code: string;
  name: string;
  created_at: string;
}